requirements.txt
flask==3.0.3 
flask-login==0.6.3 
python-dotenv==1.0.1 
flask-sqlalchemy==3.1.1 
gunicorn==23.0.0 
psycopg2-binary==2.9.10